+++ ntex-service/src/lib.rs
@@ -22,7 +22,7 @@ mod pipeline;
 mod then;
 
 pub use self::apply::{apply_fn, apply_fn_factory};
-pub use self::ctx::{Container, ContainerFactory, ServiceCall, ServiceCtx};
+pub use self::ctx::{Container, ContainerCall, ContainerFactory, ServiceCall, ServiceCtx};
 pub use self::fn_service::{fn_factory, fn_factory_with_config, fn_service};
 pub use self::fn_shutdown::fn_shutdown;
 pub use self::map_config::{map_config, unit_config};